requirements.txt
Flask==2.3.3
web3==6.11.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-binance==1.0.19
tronpy==0.2.7
solana==0.30.2